implementation of Byte Oriented Instructions

+++ picSimu/Simulation/Lib.cs
@@ -2,7 +2,7 @@ namespace picSimu.Simulation;
 
 public static class Lib
 {
-    public static bool IsBitSet(ushort b, int pos)
+    public static bool IsBitSet(uint b, int pos)
     {
         return (b & (1 << pos)) != 0;
     }
